services/news-service/pyproject.toml
[project]
name = "news-service"
version = "0.1.0"
description = "Keiko News Service"
authors = [
    {name = "Keiko Team", email = "team@keiko.ai"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "azure-cosmos>=4.8.0",
    "openai>=1.47.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
